cloud_functions/python/econ_report_by_well/requirements.txt
-r combocurve/core-requirements.txt
fpdf==1.7.2
google-cloud-bigquery-storage==2.7.0
google-cloud-bigquery[bqstorage,pandas]==2.26.0
google-cloud-storage==1.32.0
mongoengine==0.20.0
numpy==1.24.2
pandas~=1.3.5
pyarrow==5.0.0
scipy==1.10.0
xlrd==1.2.0
polars~=0.15.18
joblib~=1.2.0
